requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.5.0

# ── AI / LLM ──────────────────────────────────────────────────────────────────
openai>=1.10.0
requests>=2.31.0
langchain-openai>=0.0.5
langgraph>=0.0.26

# ── Graph / Data ───────────────────────────────────────────────────────────────
neo4j>=5.15.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
# ── UI ─────────────────────────────────────────────────────────────────────
streamlit>=1.28.0
# ── Utils ──────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
Pillow>=10.2.0
pandas>=2.1.0
